Refactoring to favor data structure
Prior to this commit, I had a data structure that "built itself" from an
internal process.

With this refactor, I created a simple data structure and a method to
extract the data structure from a given file.

The benefit of this setup is that I don't need to run the pdfimages
command against a file (well, I probably should run it at least once)
every time I want to do something/test a PDF processing.

Related to:

- https://github.com/scientist-softserv/iiif_print/issues/194
- https://github.com/samvera-labs/bulkrax/issues/760
- https://github.com/scientist-softserv/utk-hyku/issues/343
- https://github.com/scientist-softserv/adventist-dl/issues/330