Narrowing `identify` -format option for speed-up
Instead of relying on the verbose output, we're asking for the very
specific and target output of the `identify` ImageMagick command.

Prior to this commit, here are the specs for the changed file:

```shell
❯ rspec /Users/jeremy/git/space_stone-derivatives/spec/space_stone/derivatives/utilities/image_identify_spec.rb
...identify: CorruptImageProfile `/Users/jeremy/git/space_stone-derivatives/spec/fixtures/files/minimal-1-page.pdf' (XMP) @ warning/profile.c/ValidateXMPProfile/1935.
.

Finished in 4.99 seconds (files took 0.27182 seconds to load)
4 examples, 0 failures
```

With this commit, here are the specs:

```shell
❯ rspec /Users/jeremy/git/space_stone-derivatives/spec/space_stone/derivatives/utilities/image_identify_spec.rb
...identify: CorruptImageProfile `/Users/jeremy/git/space_stone-derivatives/spec/fixtures/files/minimal-1-page.pdf' (XMP) @ warning/profile.c/ValidateXMPProfile/1935.
.

Finished in 0.25737 seconds (files took 0.36577 seconds to load)
4 examples, 0 failures
```

That is a ×20 speed increase!

This is related to:

- https://github.com/scientist-softserv/iiif_print/pull/208