Incorporating mime type detection
(And a lot more)

I had a realization that the fits xml file is itself a derivative; one
that we create late in the life-cycle.  We are also running tasks that
use the technical metadata, which could be pulled from the fits xml.

Hence the "stubbing out" of the logic.  I do, however, strongly believe
that we need the mime_type.  We would generate derivatives based on what
was provided.

I could imagine a case where there would be different queues based on
the mime-type; which would allow different lambda configurations.
The design supports that option, but it is a pathway I'm not pursuing.