Renaming Type to Step
This helps with the concept of :mime_type being a derivative and having
a mime_type_step.  Right now the keyword argument of :derivative can
likely be replaced with :step; but that is a major refactor.

> There’s a conflation of Type and Step. A Type conceptual represents a
> file that can exist. A step would be responsible for building that
> file. For example the MimeType type doesn’t create a file but it does
> serve as a vitally important step in the process. And given that we
> might consider creating a FitsType, the MimeType is potentially
> dependent on that.

Note: In this process I made heavy use of two scripts:

- `find . -name "*<from>*" | sed -e "p;s/<from>/<to>/" | xargs -n2 mv`
- regular expression replacement

Once done, I ran all of the tests to ensure that things worked.  I also
ran `yard` to ensure there were no documentation generation errors.

Closes #10